changes for trust region

+++ incremental_calibration/include/aslam/calibration/algorithms/matrixOperations.tpp
@@ -50,23 +50,23 @@ namespace aslam {
       const size_t numCols = R.cols();
       const size_t dim = numCols - colBegin;
       Eigen::MatrixXd covariance = Eigen::MatrixXd::Zero(dim, dim);
-      for (int l = numCols - 1, Sigma_l = dim - 1;
-          l >= (int)(numCols - dim); --l, --Sigma_l) {
+      for (ssize_t l = numCols - 1, Sigma_l = dim - 1;
+          l >= (ssize_t)(numCols - dim); --l, --Sigma_l) {
         double temp1 = 0;
-        for (int j = l + 1, Sigma_j = Sigma_l + 1;
-            j < (int)numCols; ++j, ++Sigma_j)
+        for (ssize_t j = l + 1, Sigma_j = Sigma_l + 1;
+            j < (ssize_t)numCols; ++j, ++Sigma_j)
           temp1 += R(l, j) * covariance(Sigma_j, Sigma_l);
         const double R_ll = R(l, l);
         covariance(Sigma_l, Sigma_l) = 1 / R_ll * (1 / R_ll - temp1);
-        for (int i = l - 1, Sigma_i = Sigma_l - 1;
-            i >= int(numCols - dim); --i, --Sigma_i) {
+        for (ssize_t i = l - 1, Sigma_i = Sigma_l - 1;
+            i >= ssize_t(numCols - dim); --i, --Sigma_i) {
           temp1 = 0;
-          for (int j = i + 1, Sigma_j = Sigma_i + 1;
+          for (ssize_t j = i + 1, Sigma_j = Sigma_i + 1;
               j <= l; ++j, ++Sigma_j)
             temp1 += R(i, j) * covariance(Sigma_j, Sigma_l);
           double temp2 = 0;
-          for (int j = l + 1, Sigma_j = Sigma_l + 1; j < (int)numCols; ++j,
-              ++Sigma_j)
+          for (ssize_t j = l + 1, Sigma_j = Sigma_l + 1; j < (ssize_t)numCols;
+              ++j, ++Sigma_j)
             temp2 += R(i, j) * covariance(Sigma_l, Sigma_j);
           covariance(Sigma_i, Sigma_l) = 1 / R(i, i) * (-temp1 - temp2);
           covariance(Sigma_l, Sigma_i) = covariance(Sigma_i, Sigma_l);
